requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.1.0

# LLM and AI
google-generativeai>=0.4.0
openai>=1.12.0
langchain>=0.1.0
langchain-google-genai>=0.0.8
langchain-openai>=0.0.5

# Image Processing
Pillow>=10.0.0
PyMuPDF>=1.23.0

# OCR (fallback)
pytesseract>=0.3.10

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.26.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate Limiting
slowapi>=0.1.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Logging and Monitoring
loguru>=0.7.0
